mlts: Multilevel Latent Time Series Models with 'R' and 'Stan'

Fit multilevel manifest or latent time-series models, including popular Dynamic Structural Equation Models (DSEM). The models can be set up and modified with user-friendly functions and are fit to the data using 'Stan' for Bayesian inference. Path models and formulas for user-defined models can be easily created with functions using 'knitr'. Asparouhov, Hamaker, & Muthen (2018) doi:10.1080/10705511.2017.1406803 .

See also

Author

Maintainer: Kenneth Koslowski kenneth.koslowski@uni-leipzig.de (ORCID) [copyright holder]

Authors:

Fabian Münch fabian.felix.muench@uni-jena.de (ORCID)

Tobias Koch tobias.koch@uni-jena.de (ORCID)

Jana Holtmann jana.holtmann@uni-leipzig.de (ORCID)