Create TeX Model Formula from mlts model object

Create TeX Model Formula from mlts model object

mlts_model_formula(
  model,
  file = NULL,
  keep_tex = FALSE,
  ts = NULL,
  covariates = NULL,
  outcomes = NULL
)

Arguments

model

A model built with mlts_model.

file

An optional string containing the name of the file and file path. Has to end with .pdf file format.

keep_tex

Logical. Should the TeX file be kept (additional to the Rmd file)? Defaults to FALSE.

ts

To be included in future releases. An optional character vector containing the names of the time-series variables or indicators.

covariates

To be included in future releases. An optional character vector containing the names of the between-level covariates.

outcomes

To be included in future releases. An optional character vector containing the names of the between-level outcomes.

Value

An RMarkdown file that is automatically rendered to a pdf document.

Examples

# \donttest{
# build a simple vector-autoregressive mlts model with two time-series variables
var_model <- mlts_model(q = 2)

# create formula from the specified model
mlts_model_formula(model = var_model)
#> 
#> 
#> processing file: formula.rmd
#> 1/3        
#> 2/3 [setup]
#> 3/3        
#> output file: formula.knit.md
#> /opt/homebrew/bin/pandoc +RTS -K512m -RTS formula.knit.md --to latex --from markdown+autolink_bare_uris+tex_math_single_backslash --output formula.tex --lua-filter /Library/Frameworks/R.framework/Versions/4.5-arm64/Resources/library/rmarkdown/rmarkdown/lua/pagebreak.lua --lua-filter /Library/Frameworks/R.framework/Versions/4.5-arm64/Resources/library/rmarkdown/rmarkdown/lua/latex-div.lua --embed-resources --standalone --syntax-highlighting tango --pdf-engine pdflatex --variable graphics --extract-media formula_files --include-in-header /var/folders/z9/x7qs75yx7cd6_hwgb1f5wk9m0000gn/T//RtmpIAWpwt/rmarkdown-str13ba560ca3e63.html --variable 'geometry:margin=1in' 
#> Warning: error in running command
#> ! sh: pdflatex: command not found
#> Error: LaTeX failed to compile formula.tex. See https://yihui.org/tinytex/r/#debugging for debugging tips. See formula.log for more info.
# }