Create a summary of a fitted model with class mltsfit

Create a summary of a fitted model with class mltsfit

# S3 method for class 'mltsfit'
summary(
  object,
  priors = FALSE,
  se = FALSE,
  prob = 0.95,
  bpe = c("mean"),
  digits = 3,
  flag_signif = FALSE,
  ...
)

Arguments

object

An object of class mltsfit.

priors

Add prior information (default = FALSE).

se

Logical. Should the Monte Carlo Standard Error be included in the summary? Defaults to FALSE.

prob

A value between 0 and 1 to indicate the width of the credible interval. Default is .95.

bpe

Bayesian posterior estimate can be either "mean" (the default) or the "median" of the posterior distribution.

digits

Number of digits.

flag_signif

Add significance flags based on prob (default = FALSE).

...

Additional arguments affecting the summary produced.

Value

A summary of model parameters.

Examples

# \donttest{
# build simple vector-autoregressive mlts model for two time-series variables
var_model <- mlts_model(q = 2)

# fit model with (artificial) dataset ts_data
fit <- mlts_fit(
  model = var_model,
  data = ts_data,
  ts = c("Y1", "Y2"), # time-series variables
  id = "ID", # identifier variable
  time = "time",
  tinterval = 1 # interval for approximation of continuous-time dynamic model,
)
#> Time series variables as indicated by parameter subscripts: 
#>    1 --> Y1
#>    2 --> Y2
#> 
#>  Note that obtaining (standardized) estimates by cluster,
#>     requires settting `monitor_person_pars = TRUE`. However, keeping the default option
#>     (`monitor_person_pars = FALSE`) can improve sampling times.
#> Warning: There were 7 divergent transitions after warmup. See
#> https://mc-stan.org/misc/warnings.html#divergent-transitions-after-warmup
#> to find out why this is a problem and how to eliminate them.
#> Warning: There were 1 chains where the estimated Bayesian Fraction of Missing Information was low. See
#> https://mc-stan.org/misc/warnings.html#bfmi-low
#> Warning: Examine the pairs() plot to diagnose sampling problems
#> Warning: The largest R-hat is NA, indicating chains have not mixed.
#> Running the chains for more iterations may help. See
#> https://mc-stan.org/misc/warnings.html#r-hat
#> Warning: Bulk Effective Samples Size (ESS) is too low, indicating posterior means and medians may be unreliable.
#> Running the chains for more iterations may help. See
#> https://mc-stan.org/misc/warnings.html#bulk-ess
#> Warning: Tail Effective Samples Size (ESS) is too low, indicating posterior variances and tail quantiles may be unreliable.
#> Running the chains for more iterations may help. See
#> https://mc-stan.org/misc/warnings.html#tail-ess

# inspect model summary
summary(fit)
#> Time series variables as indicated by parameter subscripts: 
#>    1 --> Y1
#>    2 --> Y2
#> Data: 1100 observations in 55 IDs
#> Model convergence criteria: 
#>   Maximum Potential Scale Reduction Factor (PSR; Rhat): 1.208 (should be < 1.01)
#>   Minimum Bulk ESS: 7 (should be > 200, 100 per chain) 
#>   Minimum Tail ESS: 8 (should be > 200, 100 per chain) 
#>   Number of divergent transitions: 7 (should be 0) 
#> 
#> Posterior Summary Statistics
#> Fixed Effects:
#>                Mean    SD   2.5%  97.5%  Rhat Bulk_ESS Tail_ESS
#>         mu_1  0.791 0.132  0.505  1.052 1.015      107      350
#>         mu_2  0.336 0.172 -0.034  0.654 1.019      127      324
#>    phi(1)_11  0.347 0.040  0.272  0.432 1.045       44      141
#>    phi(1)_12 -0.114 0.045 -0.198 -0.026 1.022      129      361
#>    phi(1)_21  0.056 0.039 -0.022  0.131 1.065      216      324
#>    phi(1)_22  0.134 0.043  0.056  0.214 1.069       21       10
#>  ln.sigma2_1 -0.220 0.061 -0.310 -0.105 1.127       12        8
#>  ln.sigma2_2 -0.293 0.063 -0.430 -0.181 1.036       71      305
#>    r.zeta_12 -0.149 0.031 -0.205 -0.085 1.013      192      348
#> 
#> Random Effects SDs:
#>                     Mean    SD  2.5% 97.5%  Rhat Bulk_ESS Tail_ESS
#>         sigma_mu_1 1.034 0.108 0.820 1.260 1.018      339      274
#>         sigma_mu_2 1.302 0.127 1.093 1.570 1.020       74      331
#>    sigma_phi(1)_11 0.113 0.048 0.020 0.198 1.208        7       18
#>    sigma_phi(1)_12 0.243 0.051 0.141 0.349 1.035       67       64
#>    sigma_phi(1)_21 0.186 0.051 0.082 0.276 1.121       12       83
#>    sigma_phi(1)_22 0.188 0.044 0.094 0.275 1.021       79      123
#>  sigma_ln.sigma2_1 0.161 0.079 0.023 0.314 1.146       15       23
#>  sigma_ln.sigma2_2 0.298 0.066 0.183 0.442 1.089       50      242
#> 
#> Random Effects Correlations:
#>                            Mean    SD   2.5%  97.5%  Rhat Bulk_ESS Tail_ESS
#>                mu_1.mu_2 -0.125 0.137 -0.377  0.158 1.015      133      353
#>           mu_1.phi(1)_11 -0.053 0.244 -0.486  0.515 1.053      190      231
#>           mu_1.phi(1)_12  0.264 0.187 -0.082  0.588 1.018      118       65
#>           mu_1.phi(1)_21 -0.085 0.202 -0.427  0.343 1.031       70      286
#>           mu_1.phi(1)_22 -0.125 0.201 -0.539  0.248 1.038      114      211
#>         mu_1.ln.sigma2_1  0.077 0.244 -0.404  0.538 1.119      248      223
#>         mu_1.ln.sigma2_2  0.182 0.221 -0.188  0.577 1.057       26       11
#>           mu_2.phi(1)_11 -0.165 0.254 -0.633  0.322 1.011      121      253
#>           mu_2.phi(1)_12  0.062 0.156 -0.278  0.372 1.010      345      288
#>           mu_2.phi(1)_21  0.115 0.209 -0.269  0.499 1.002      133       51
#>           mu_2.phi(1)_22 -0.401 0.190 -0.754 -0.034 1.044       42      213
#>         mu_2.ln.sigma2_1 -0.094 0.258 -0.575  0.402 1.172      115       71
#>         mu_2.ln.sigma2_2 -0.030 0.175 -0.398  0.255 1.005      300      502
#>      phi(1)_11.phi(1)_12 -0.017 0.321 -0.530  0.592 1.142       11       51
#>      phi(1)_11.phi(1)_21  0.185 0.331 -0.428  0.875 1.087       30       23
#>      phi(1)_11.phi(1)_22  0.212 0.312 -0.277  0.740 1.111       13       14
#>    phi(1)_11.ln.sigma2_1  0.194 0.315 -0.477  0.740 1.047       40      136
#>    phi(1)_11.ln.sigma2_2 -0.008 0.280 -0.572  0.525 1.059       67      157
#>      phi(1)_12.phi(1)_21 -0.219 0.230 -0.634  0.191 1.006      122      273
#>      phi(1)_12.phi(1)_22  0.155 0.238 -0.358  0.608 1.003      181      303
#>    phi(1)_12.ln.sigma2_1  0.308 0.272 -0.269  0.749 1.022       84      270
#>    phi(1)_12.ln.sigma2_2  0.055 0.261 -0.457  0.500 1.023       63      201
#>      phi(1)_21.phi(1)_22  0.177 0.228 -0.320  0.570 1.019      100      229
#>    phi(1)_21.ln.sigma2_1 -0.093 0.285 -0.593  0.514 1.032      100       70
#>    phi(1)_21.ln.sigma2_2 -0.255 0.231 -0.726  0.241 1.068       72      187
#>    phi(1)_22.ln.sigma2_1  0.317 0.249 -0.230  0.722 1.049       79      185
#>    phi(1)_22.ln.sigma2_2 -0.097 0.252 -0.489  0.362 1.061       28      176
#>  ln.sigma2_1.ln.sigma2_2 -0.154 0.260 -0.643  0.445 1.055      130      235
#> 
#> Samples were drawn using NUTS on Wed Jun 17 16:55:03 2026.
#> For each parameter, Bulk_ESS and Tail_ESS are measures of effective
#> sample size, and Rhat is the potential scale reduction factor
#> on split chains (at convergence, Rhat = 1).
# }