Fit Bayesian Multilevel Manifest or Latent Time-Series Models

Fit Bayesian Multilevel Manifest or Latent Time-Series Models

mlts_fit(
  model,
  data = NULL,
  id,
  group = NULL,
  ts,
  covariates = NULL,
  outcomes = NULL,
  outcome_pred_btw = NULL,
  center_covs = TRUE,
  time = NULL,
  tinterval = NULL,
  beep = NULL,
  days = NULL,
  n_overnight_NAs,
  max_NA_seq = NULL,
  na.rm = FALSE,
  iter = 500,
  chains = 2,
  cores = 2,
  monitor_person_pars = FALSE,
  monitor_all_pars = FALSE,
  get_SD_latent = FALSE,
  fit_model = TRUE,
  print_message = TRUE,
  print_warning = TRUE,
  ...
)

Arguments

model

data.frame. Output of mlts_model and related functions.

data

An object of class data.frame (or one that can be coerced to that class) containing data of all variables used in the model. Alternatively, a list object with simulated data created by mlts_sim can be entered directly and allows for comparison of estimates and true population paramter values used in the data generation.

id

Character. The variable in data that identifies the observational cluster unit. Not necessary when data is a list object of simulated data generated with mlts_sim.

group

Character. The variable in data that identifies the grouping variable across cluster units.

ts

Character. The variable(s) in data that contain the time-series construct(s) or their indicator variable(s). If multiple constructs are provided in the model, multiple entries are necessary. Note that the order of variable names provided in ts has to match the specification made in the model. E.g., if multiple constructs (e.g., mlts_model(q = 2)) are provided the order of variables names provided in ts determines which construct is referred to as mu_1, phi(1)_11, etc..

covariates

Named character vector. An optional named vector of characters to refer to predictors of random effects as specified in the model. Note that specifying covariates is only necessary if the respective variable name(s) in data differ from the variables names specified in model.

outcomes

Named character vector. Similar to covariates, an optional named vector of characters to refer to outcome predicted by random effects as specified in the model. Note that specifying outcomes is only necessary if the respective variable name(s) in data differ from the outcome variable name(s) specified in model.

outcome_pred_btw

Named character vector. Similar to covariates, an optional named vector of characters to refer to additional between-level variables entered as outcome predictor(s) as specified in the model. Note that specifying outcome_pred_btw is only necessary if the respective variable name(s) in data differ from the variable name(s) specified in model.

center_covs

Logical. Between-level covariates used as predictors of random effects will be grand-mean centered before model fitting by default. Set center_covs to FALSE when including categorical predictors into the set of covariates. Note that in this case, additional continuous covariates should be grand-mean centered prior to using mlts_fit. If group is specified, covariates will be centered on their respective group mean.

time

Character. The variable in data that contains the (continuous) time of observation.

tinterval

The step interval for approximating equally spaced observations in time by insertion of missing values, to be specified with respect to the time stamp variable provided in time. Procedure for inserting missing values resembles the procedure for time shift transformation as described in Asparouhov, Hamaker, & Muthén (2018).

beep

Character. The variable in data that contains the running beep number starting with 1 for each person.

days

Optional. If a running beep identifier is provided via the beep argument and observations are nested within days (or similar grouping unit), the variable in data that contains the day identifier can be added to correct for overnight lags (see Details).

n_overnight_NAs

Optional. The number of NA rows to add after the last observation of each day (if days is provided).

max_NA_seq

Integer. Specify a maximum number of consecutive missing values. Can decrease estimation times drastically in the presence of very long sequences of missing values (e.g., when setting tinterval to values of small time steps).

na.rm

logical. Per default, missing values remain in the data and will be imputed during model estimation. Set to TRUE to remove all rows with missing values in variables given in ts.

iter

A positive integer specifying the number of iterations for each chain (including 50% used as warmup). The default is 500.

chains

A positive integer specifying the number of Markov chains. The default is 2.

cores

The number of cores to use when executing the Markov chains in parallel. The default is 2 (see stan).

monitor_person_pars

Logical. Should person parameters (i.e., values of the latent variables) be stored? Default is FALSE.

monitor_all_pars

Logical. Should all parameters be stored? Default is FALSE.

get_SD_latent

Logical. Set to TRUE to obtain standardized estimates in multiple-indicator models.

fit_model

Logical. Set to FALSE to avoid fitting the model which may be helpful to inspect prepared data used for model estimation (default = TRUE).

print_message

Logical. Print messages based on defined inputs (default = TRUE).

print_warning

Logical. Print warnings based on defined inputs (default = TRUE).

...

Additional arguments passed to sampling.

Value

An object of class mltsfit. The object is a list containing the following components:

model

the model object passed to mlts_fit

data

the preprocessed data used for fitting the model

param.labels

a data.frame that provides the names of parameters used in the stan model. These parameter names are necessary when running standard post-processing functions using mlts_fit$stanfit

pop.pars.summary

a data.frame that contains summary statistics for all parameter in model

person.pars.summary

if monitor_person_pars = TRUE, a data.frame containing summary statistics for cluster-specific parameters is provided

standata

a list with the data as passed to sampling

stanfit

an object of class stanfit with the raw output created by sampling

posteriors

an array of the MCMC chain results for all parameters in model created by rstan::extract with dimnames adapted to match the parameter names provided in model

References

Asparouhov, T., Hamaker, E. L., & Muthén, B. (2018). Dynamic Structural Equation Models. Structural Equation Modeling: A Multidisciplinary Journal, 25(3), 359–388. doi:10.1080/10705511.2017.1406803

Examples

# \donttest{
# build simple vector-autoregressive mlts model for two time-series variables
var_model <- mlts_model(q = 2)

# fit model with (artificial) dataset ts_data
fit <- mlts_fit(
  model = var_model,
  data = ts_data,
  ts = c("Y1", "Y2"), # time-series variables
  id = "ID", # cluster identifier variable
  time = "time", # time variable
  tinterval = 1 # interval for approximation of equidistant measurements,
)
#> Time series variables as indicated by parameter subscripts: 
#>    1 --> Y1
#>    2 --> Y2
#> 
#>  Note that obtaining (standardized) estimates by cluster,
#>     requires settting `monitor_person_pars = TRUE`. However, keeping the default option
#>     (`monitor_person_pars = FALSE`) can improve sampling times.
#> Warning: There were 3 divergent transitions after warmup. See
#> https://mc-stan.org/misc/warnings.html#divergent-transitions-after-warmup
#> to find out why this is a problem and how to eliminate them.
#> Warning: There were 2 chains where the estimated Bayesian Fraction of Missing Information was low. See
#> https://mc-stan.org/misc/warnings.html#bfmi-low
#> Warning: Examine the pairs() plot to diagnose sampling problems
#> Warning: The largest R-hat is NA, indicating chains have not mixed.
#> Running the chains for more iterations may help. See
#> https://mc-stan.org/misc/warnings.html#r-hat
#> Warning: Bulk Effective Samples Size (ESS) is too low, indicating posterior means and medians may be unreliable.
#> Running the chains for more iterations may help. See
#> https://mc-stan.org/misc/warnings.html#bulk-ess
#> Warning: Tail Effective Samples Size (ESS) is too low, indicating posterior variances and tail quantiles may be unreliable.
#> Running the chains for more iterations may help. See
#> https://mc-stan.org/misc/warnings.html#tail-ess

# inspect model summary
summary(fit)
#> Time series variables as indicated by parameter subscripts: 
#>    1 --> Y1
#>    2 --> Y2
#> Data: 1100 observations in 55 IDs
#> Model convergence criteria: 
#>   Maximum Potential Scale Reduction Factor (PSR; Rhat): 1.254 (should be < 1.01)
#>   Minimum Bulk ESS: 9 (should be > 200, 100 per chain) 
#>   Minimum Tail ESS: 12 (should be > 200, 100 per chain) 
#>   Number of divergent transitions: 3 (should be 0) 
#> 
#> Posterior Summary Statistics
#> Fixed Effects:
#>                Mean    SD   2.5%  97.5%  Rhat Bulk_ESS Tail_ESS
#>         mu_1  0.782 0.153  0.486  1.079 1.013      855      313
#>         mu_2  0.305 0.185 -0.088  0.637 1.014      428      210
#>    phi(1)_11  0.348 0.040  0.264  0.424 1.006      179      332
#>    phi(1)_12 -0.115 0.047 -0.207 -0.022 1.010      315      248
#>    phi(1)_21  0.056 0.038 -0.014  0.132 1.004      338      318
#>    phi(1)_22  0.138 0.040  0.062  0.215 1.015      198      367
#>  ln.sigma2_1 -0.214 0.059 -0.333 -0.097 1.045       46      224
#>  ln.sigma2_2 -0.290 0.058 -0.413 -0.185 1.015      205      331
#>    r.zeta_12 -0.148 0.033 -0.210 -0.084 1.007      554      355
#> 
#> Random Effects SDs:
#>                     Mean    SD  2.5% 97.5%  Rhat Bulk_ESS Tail_ESS
#>         sigma_mu_1 1.029 0.124 0.810 1.293 1.008      462      207
#>         sigma_mu_2 1.291 0.137 1.065 1.589 0.998      776      371
#>    sigma_phi(1)_11 0.123 0.048 0.035 0.214 1.076       25      119
#>    sigma_phi(1)_12 0.238 0.050 0.143 0.339 1.003      174      231
#>    sigma_phi(1)_21 0.157 0.072 0.009 0.273 1.254        9       12
#>    sigma_phi(1)_22 0.192 0.047 0.114 0.287 1.010       78      175
#>  sigma_ln.sigma2_1 0.200 0.074 0.064 0.351 1.045       37       77
#>  sigma_ln.sigma2_2 0.257 0.083 0.094 0.416 1.070       39      110
#> 
#> Random Effects Correlations:
#>                            Mean    SD   2.5%  97.5%  Rhat Bulk_ESS Tail_ESS
#>                mu_1.mu_2 -0.111 0.131 -0.365  0.128 1.005      530      391
#>           mu_1.phi(1)_11 -0.054 0.254 -0.567  0.440 1.018      264      242
#>           mu_1.phi(1)_12  0.220 0.166 -0.118  0.555 1.009      275      379
#>           mu_1.phi(1)_21 -0.083 0.247 -0.663  0.363 1.034      194       58
#>           mu_1.phi(1)_22 -0.118 0.203 -0.475  0.291 1.002      296      329
#>         mu_1.ln.sigma2_1  0.081 0.243 -0.368  0.540 1.026      209      337
#>         mu_1.ln.sigma2_2  0.256 0.220 -0.169  0.660 1.000      276      331
#>           mu_2.phi(1)_11 -0.153 0.256 -0.620  0.361 1.008      217      288
#>           mu_2.phi(1)_12  0.080 0.171 -0.258  0.393 0.998      464      477
#>           mu_2.phi(1)_21  0.120 0.206 -0.283  0.507 1.004      444      338
#>           mu_2.phi(1)_22 -0.408 0.182 -0.734 -0.053 1.016      107      116
#>         mu_2.ln.sigma2_1 -0.063 0.234 -0.476  0.428 1.004      317      314
#>         mu_2.ln.sigma2_2 -0.052 0.206 -0.438  0.366 1.008      314      432
#>      phi(1)_11.phi(1)_12  0.072 0.258 -0.455  0.577 1.008      213      245
#>      phi(1)_11.phi(1)_21  0.163 0.312 -0.439  0.714 1.019      107       86
#>      phi(1)_11.phi(1)_22  0.278 0.280 -0.274  0.751 1.047       73      184
#>    phi(1)_11.ln.sigma2_1  0.270 0.302 -0.385  0.798 1.114       15      108
#>    phi(1)_11.ln.sigma2_2 -0.085 0.295 -0.639  0.506 1.002      132      235
#>      phi(1)_12.phi(1)_21 -0.233 0.263 -0.692  0.309 1.055       40       64
#>      phi(1)_12.phi(1)_22  0.110 0.261 -0.489  0.573 1.009       83       61
#>    phi(1)_12.ln.sigma2_1  0.353 0.243 -0.157  0.759 1.017      158      199
#>    phi(1)_12.ln.sigma2_2 -0.002 0.275 -0.519  0.534 1.031       68       50
#>      phi(1)_21.phi(1)_22  0.099 0.260 -0.422  0.557 1.024       78      313
#>    phi(1)_21.ln.sigma2_1 -0.095 0.285 -0.629  0.478 1.009      161      191
#>    phi(1)_21.ln.sigma2_2 -0.185 0.285 -0.669  0.434 1.070       25       46
#>    phi(1)_22.ln.sigma2_1  0.346 0.259 -0.306  0.747 1.055       44      358
#>    phi(1)_22.ln.sigma2_2 -0.049 0.267 -0.523  0.522 1.030      229      183
#>  ln.sigma2_1.ln.sigma2_2 -0.132 0.286 -0.659  0.426 1.002      179      319
#> 
#> Samples were drawn using NUTS on Wed Jun 17 16:53:05 2026.
#> For each parameter, Bulk_ESS and Tail_ESS are measures of effective
#> sample size, and Rhat is the potential scale reduction factor
#> on split chains (at convergence, Rhat = 1).
# }